Revert "Merge pull request #101 from appliedAI-Initiative/93-move-documentation-of-influence-functions-to-docs-instead-of-readme"
This reverts commit b0d2ab2f2f8cf6277c50a7089db1f8643d55287c, reversing
changes made to 08c8be95fee9d61a588ad9fb5b00dbd177cb94e1.
