--- src/valuation/influence/__init__.py
@@ -1,3 +1,0 @@
-"""
-Contains influence function calculation functions.
-"""
