+++ src/valuation/solve/cg.py
@@ -1,12 +1,5 @@
-"""
-Contains
-
-- batched conjugate gradient.
-- error bound for conjugate gradient.
-"""
-
 from types import LambdaType
-from typing import Callable, Optional, Union
+from typing import Callable, Union
 
 import numpy as np
 
@@ -15,34 +8,30 @@ from valuation.utils import (
     mcmc_is_linear_function_positive_definite,
 )
 from valuation.utils.logging import raise_or_log
-from valuation.utils.types import MatrixVectorProduct
 
 
-def batched_preconditioned_conjugate_gradient(
+def conjugate_gradient(
     A: Union[np.ndarray, Callable[[np.ndarray], np.ndarray]],
     b: np.ndarray,
     x0: np.ndarray = None,
     M: Union[np.ndarray, Callable[[np.ndarray], np.ndarray]] = None,
-    rtol: float = 1e-10,
+    rtol: float = 1e-5,
     max_iterations: int = None,
-    max_step_size: Optional[float] = 10.0,
     verify_assumptions: bool = False,
     raise_exception: bool = False,
 ):
     """
-    Implementation of a batched conjugate gradient algorithm. It uses vector matrix products for efficient calculation.
-    See https://en.wikipedia.org/wiki/Conjugate_gradient_method for more details of the algorithm. See also
+    Implementation of a batched conjugate gradient algorithm. It uses vector matrix products for efficient calculation. See
+    https://en.wikipedia.org/wiki/Conjugate_gradient_method for more details of the algorithm. See also
     https://github.com/scipy/scipy/blob/v1.8.1/scipy/sparse/linalg/_isolve/iterative.py#L282-L351 and
-    https://web.stanford.edu/class/ee364b/lectures/conj_grad_slides.pdf. On top, it constrains the maximum step size.
+    https://web.stanford.edu/class/ee364b/lectures/conj_grad_slides.pdf.
     :param A: A linear function f : R[k] -> R[k] representing a matrix vector product from dimension K to K or a matrix.
     It has to be positive-definite v.T @ f(v) >= 0.
     :param b: A np.ndarray of shape [K] representing the targeted result of the matrix multiplication Ax.
     :param M: A function f : R[k] -> R[k] which approximates inv(A) or a matrix of shape [K, K]. The underlying matrix
     has to be symmetric and positive definite.
-    :param max_iterations: Maximum number of iterations to use in conjugate gradient. Default is 10 times K.
+    :param max_iterations: The maximum number of iterations to use in conjugate gradient.
     :param rtol: Relative tolerance of the residual with respect to the 2-norm of b.
-    :param max_step_size: Maximum step size along a gradient direction. Might be necessary for numerical stability.
-    See also max_iterations. Default is 10.0.
     :param verify_assumptions: True, iff the matrix should be checked for positive-definiteness by a stochastic rule.
     :param raise_exception: True, iff an assumption should be raised, instead of a warning only.
     :return: A np.ndarray of shape [K] representing the solution of Ax=b.
@@ -113,9 +102,7 @@ def batched_preconditioned_conjugate_gradient(
         r_dot_u = np.einsum(
             "ia,ia->i", r[not_yet_converged_indices], u[not_yet_converged_indices]
         )
-        pre_alpha = r_dot_u / p_dot_mvp
-        if max_step_size is not None:
-            alpha = np.minimum(max_step_size, pre_alpha)
+        alpha = r_dot_u / p_dot_mvp
 
         # update x and r
         reshaped_alpha = alpha.reshape([-1, 1])
@@ -159,18 +146,11 @@ def batched_preconditioned_conjugate_gradient(
     return x, iteration
 
 
-def conjugate_gradient_condition_number_based_error_bound(
+def conjugate_gradient_error_bound(
     A: np.ndarray, n: int, x0: np.ndarray, xt: np.ndarray
 ) -> float:
     """
-    Error bound for conjugate gradient based on the condition number of the weight matrix A. Used for testing purposes.
-    See also https://math.stackexchange.com/questions/382958/error-for-conjugate-gradient-method. Explicit of the weight
-    matrix is required.
-    :param A: Weight matrix of the matrix to be inverted.
-    :param n: Maximum number for executed iterations X in conjugate gradient.
-    :param x0: Initialization solution x0 of conjugate gradient.
-    :param xt: Final solution xt of conjugate gradient after X iterations.
-    :returns: Upper bound for ||x0 - xt||_A.
+    https://math.stackexchange.com/questions/382958/error-for-conjugate-gradient-method
     """
     eigvals = np.linalg.eigvals(A)
     eigvals = np.sort(eigvals)
@@ -184,26 +164,3 @@ def conjugate_gradient_condition_number_based_error_bound(
     div = (sqrt_kappa + 1) / (sqrt_kappa - 1)
     div_n = div**n
     return (2 * error_init) / (div_n + 1 / div_n)
-
-
-def hvp_to_inv_diag_conditioner(
-    hvp: MatrixVectorProduct, d: int
-) -> MatrixVectorProduct:
-    """
-    This method uses the hvp function to construct a simple pre-conditioner 1/diag(H). It does so while requiring
-    only O(d) space in RAM for construction and later execution.
-    :param hvp: The callable calculating the Hessian vector product Hv.
-    :param d: The number of dimensions of the hvp callable.
-    :returns: A MatrixVectorProduct for the conditioner.
-    """
-    diags = np.empty(d)
-
-    for i in range(d):
-        inp = np.zeros(d)
-        inp[i] = 1
-        diags[i] = hvp(np.reshape(inp, [1, -1]))[0, i]
-
-    def _inv_diag_conditioner(v: np.ndarray):
-        return v / diags
-
-    return _inv_diag_conditioner
