+++ tests/conftest.py
@@ -11,7 +11,7 @@ import pytest
 from pymemcache.client import Client
 from sklearn.linear_model import LinearRegression
 from sklearn.pipeline import make_pipeline
-from sklearn.preprocessing import MinMaxScaler, PolynomialFeatures
+from sklearn.preprocessing import PolynomialFeatures
 
 from valuation.utils import ClientConfig, Dataset, Utility
 from valuation.utils.logging import start_logging_server
@@ -634,10 +634,4 @@ def create_mock_dataset(
         dataset.x_test = np.random.uniform(size=[test_set_size, i_d])
         dataset.y_test = data_model(dataset.x_test)
 
-    scaler_x = MinMaxScaler()
-    scaler_y = MinMaxScaler()
-    dataset.x_train = scaler_x.fit_transform(dataset.x_train)
-    dataset.y_train = scaler_y.fit_transform(dataset.y_train)
-    dataset.x_test = scaler_x.transform(dataset.x_test)
-    dataset.y_test = scaler_y.transform(dataset.y_test)
     return dataset
