+++ tests/influence/test_conjugate_gradients.py
@@ -4,10 +4,7 @@ from typing import List, Tuple
 import numpy as np
 import pytest
 
-from valuation.solve.cg import (
-    batched_preconditioned_conjugate_gradient,
-    conjugate_gradient_condition_number_based_error_bound,
-)
+from valuation.solve.cg import conjugate_gradient, conjugate_gradient_error_bound
 
 
 class AlgorithmTestSettings:
@@ -48,7 +45,7 @@ test_case_ids = list(map(lmb_test_case_to_str, zip(range(len(test_cases)), test_
 def test_conjugate_gradients_mvp(quadratic_linear_equation_system):
     A, b = quadratic_linear_equation_system
     x0 = np.zeros_like(b)
-    xn, n = batched_preconditioned_conjugate_gradient(A, b, x0=x0)
+    xn, n = conjugate_gradient(A, b, x0=x0)
     check_solution(A, b, n, x0, xn)
 
 
@@ -63,7 +60,7 @@ def test_conjugate_gradients_fn(quadratic_linear_equation_system):
     new_A = np.copy(A)
     A = lambda v: v @ new_A.T
     x0 = np.zeros_like(b)
-    xn, n = batched_preconditioned_conjugate_gradient(A, b, x0=x0)
+    xn, n = conjugate_gradient(A, b, x0=x0)
     check_solution(new_A, b, n, x0, xn)
 
 
@@ -77,14 +74,13 @@ def test_conjugate_gradients_mvp_preconditioned(quadratic_linear_equation_system
     A, b = quadratic_linear_equation_system
     M = np.diag(1 / np.diag(A))
     x0 = np.zeros_like(b)
-    xn, n = batched_preconditioned_conjugate_gradient(A, b, M=M, x0=x0)
+    xn, n = conjugate_gradient(A, b, M=M, x0=x0)
     check_solution(A, b, n, x0, xn)
 
 
 @pytest.mark.parametrize(
     "problem_dimension,batch_size,condition_number",
     test_cases,
-    ids=test_case_ids,
     indirect=True,
 )
 def test_conjugate_gradients_singular_matrix(
@@ -93,7 +89,7 @@ def test_conjugate_gradients_singular_matrix(
     A, b = singular_quadratic_linear_equation_system
     x0 = np.zeros_like(b)
     try:
-        xn, n = batched_preconditioned_conjugate_gradient(
+        xn, n = conjugate_gradient(
             A, b, x0=x0, verify_assumptions=True, raise_exception=True
         )
         pytest.fail("Should throw an excepction.")
@@ -114,7 +110,7 @@ def check_solution(A, b, n, x0, xn):
     inv_A = np.linalg.pinv(A)
 
     xt = b @ inv_A.T
-    bound = conjugate_gradient_condition_number_based_error_bound(A, n, x0, xt)
+    bound = conjugate_gradient_error_bound(A, n, x0, xt)
     norm_A = lambda v: np.sqrt(np.einsum("ia,ab,ib->i", v, A, v))
     assert np.all(np.logical_not(np.isnan(xn)))
