+++ tests/models/test_models.py
@@ -1,9 +1,3 @@
-"""
-Contains tests for LinearRegression, BinaryLogisticRegression as well as TorchModule, TwiceDifferentiable interface and
-its associated gradient and matrix vector product calculations. Note that there is no test for the neural network
-module.
-"""
-
 import itertools
 from typing import List, Tuple
 
@@ -19,12 +13,11 @@ from valuation.utils import (
 try:
     import torch.nn.functional as F
 
-    from valuation.models import (
+    from valuation.models.binary_logistic_regression import (
         BinaryLogisticRegressionTorchModel,
-        LinearRegressionTorchModel,
-        TorchModule,
-        TorchOptimizer,
     )
+    from valuation.models.linear_regression_torch_model import LRTorchModel
+    from valuation.models.pytorch_model import PyTorchOptimizer, PyTorchSupervisedModel
 except ImportError:
     pass
 
@@ -130,12 +123,12 @@ def test_logistic_regression_model_fit(
     ).astype(int)
     train_x = np.random.uniform(low=-1, high=1, size=[train_set_size, input_dimension])
     train_y = data_model(train_x)
-    model = TorchModule(
+    model = PyTorchSupervisedModel(
         model=BinaryLogisticRegressionTorchModel(input_dimension),
         objective=F.binary_cross_entropy,
         num_epochs=1000,
         batch_size=32,
-        optimizer=TorchOptimizer.ADAM,
+        optimizer=PyTorchOptimizer.ADAM,
         optimizer_kwargs={"lr": 0.001, "weight_decay": 1e-4},
     )
     model.fit(train_x, train_y)
@@ -174,12 +167,12 @@ def test_linear_regression_model_fit(
     train_x = np.random.uniform(size=[train_set_size, input_dimension])
     train_y = data_model(train_x)
 
-    model = TorchModule(
-        model=LinearRegressionTorchModel((output_dimension, input_dimension)),
+    model = PyTorchSupervisedModel(
+        model=LRTorchModel((output_dimension, input_dimension)),
         objective=F.mse_loss,
         num_epochs=1000,
         batch_size=32,
-        optimizer=TorchOptimizer.ADAM_W,
+        optimizer=PyTorchOptimizer.ADAM_W,
         optimizer_kwargs={"lr": 0.05},
     )
     model.fit(train_x, train_y)
@@ -218,10 +211,8 @@ def test_linear_regression_model_grad(
     train_x = np.random.uniform(size=[train_set_size, input_dimension])
     train_y = data_model(train_x)
 
-    model = TorchModule(
-        model=LinearRegressionTorchModel(
-            dim=(input_dimension, output_dimension), init=linear_model
-        ),
+    model = PyTorchSupervisedModel(
+        model=LRTorchModel(dim=(input_dimension, output_dimension), init=linear_model),
         objective=F.mse_loss,
     )
 
@@ -257,10 +248,8 @@ def test_linear_regression_model_hessian(
     train_x = np.random.uniform(size=[train_set_size, input_dimension])
     train_y = data_model(train_x)
 
-    model = TorchModule(
-        model=LinearRegressionTorchModel(
-            dim=(input_dimension, output_dimension), init=linear_model
-        ),
+    model = PyTorchSupervisedModel(
+        model=LRTorchModel(dim=(input_dimension, output_dimension), init=linear_model),
         objective=F.mse_loss,
     )
 
@@ -298,14 +287,12 @@ def test_linear_regression_model_d_x_d_theta(
     train_x = np.random.uniform(size=[train_set_size, input_dimension])
     train_y = data_model(train_x)
 
-    model = TorchModule(
-        model=LinearRegressionTorchModel(
-            dim=(input_dimension, output_dimension), init=(A, b)
-        ),
+    model = PyTorchSupervisedModel(
+        model=LRTorchModel(dim=(input_dimension, output_dimension), init=(A, b)),
         objective=F.mse_loss,
         num_epochs=1000,
         batch_size=32,
-        optimizer=TorchOptimizer.ADAM_W,
+        optimizer=PyTorchOptimizer.ADAM_W,
         optimizer_kwargs={"lr": 0.02},
     )
