+++ tox.ini
@@ -53,8 +53,6 @@ deps =
 description = Tests notebooks
 commands =
     pytest notebooks
-extras:
-    torch
 
 [testenv:linting]
 skip_install = true
@@ -70,8 +68,6 @@ deps =
     pylint-json2html
     black
     isort
-extras:
-    torch
 whitelist_externals =
     bash
